+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+---------------------
+##### Fixed
+- Date format shown to user in readonly field is different from non-readonly field.
+
 [4.47.0] - 2018-02-21
 ---------------------
 ##### Added